+++ web-admin/src/components/borrow/jev/JevTraceDrawer.vue
@@ -10,6 +10,7 @@
     <template v-else>
       <div class="jev-trace-meta">
         <el-tag :type="statusTagType" effect="plain">{{ statusLabel }}</el-tag>
+        <el-tag v-if="trace.meta.force_refresh" type="warning" effect="plain">强制外呼</el-tag>
         <span v-if="trace.meta.model" class="meta-model">{{ trace.meta.model }}</span>
         <span v-if="trace.meta.error_message" class="meta-error">{{ trace.meta.error_message }}</span>
       </div>